+++ changelog.md
@@ -1,11 +1,17 @@
 # Changelog
 
+## v0.8.0
+
+- Add `subscribe.for(actions, props, state)` function.
+- Add `subscribe.props(props, state)` function.
+- Add `subscribe.actions(actions, state)` function.
+
 ## v0.7.2
 
 - Add `setBaseURL()` to set the `fetch()` base url when no `url` given (or `url = ''`), so the request url will
   become `BASE_URL/{location.pathname}{location.search}`.
 
-## 0.7.1
+## v0.7.1
 
 - Fix missing `watch()` typings.